server/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.1.1
rq==1.16.2
sqlparse==0.5.1
tzdata==2024.1
